requirements.txt
customtkinter>=5.2.2
pillow>=10.2.0
opencv-python>=4.9.0
numpy>=1.26.3
pyautogui>=0.9.54
pandas>=2.2.0
openpyxl>=3.1.2
pyinstaller>=6.3.0
psutil>=5.9.8
requests>=2.31.0
schedule>=1.2.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
pywin32>=306; sys_platform == 'win32'
